Fix empty file uploads (#78)
* Fix empty file uploads

tus-js-client and tusd support uploads of empty files. In
these cases, only the initial POST request needs to be performed:

https://github.com/tus/tus-js-client/issues/106

This was supported by 0.2.5, but the refactoring inadvertently broke
this by skipping by all HTTP requests entirely.

* Add tests for empty file uploads

* Refactor "empty file" tests

It's more reliable and concise to create an uploader from the existing
client fixture instead of modifying an existing uploader to resemble
what we want.